refactor: 전체적인 container.js 추가, 일부 css조정
