+++ pages/_document.js
@@ -6,12 +6,9 @@ class MyDocument extends Document {
     return (
       <Html>
         <Head>
-          {/* Google Fonts 링크 추가 */}
-          <link rel="preconnect" href="https://fonts.googleapis.com" />
-          <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
           <link
-            href="https://fonts.googleapis.com/css2?family=Noto+Sans+KR:wght@100..900&display=swap"
             rel="stylesheet"
+            href="https://cdn.jsdelivr.net/npm/pretendard/dist/web/static/pretendard.css"
           />
         </Head>
         <body>
