+++ pages/login/signup/01/index.js
@@ -1,0 +1,9 @@
+import SignUp01 from "../../../../src/components/units/signup/signup01/Signup01.container";
+
+export default function SignUp01Page(){
+  return(
+    <>
+      <SignUp01 />
+    </>
+  )
+}
