+++ pages/login/signup/02/index.js
@@ -1,0 +1,9 @@
+import SignUp02 from "../../../../src/components/units/signup/signup02/Signup02.container";
+
+export default function SignUp02Page(){
+  return(
+    <>
+      <SignUp02 />
+    </>
+  )
+}
