+++ pages/login/signup/03/index.js
@@ -1,0 +1,9 @@
+import SignUp03 from "../../../../src/components/units/signup/signup03/Signup03.container";
+
+export default function SignUp03Page(){
+  return(
+    <>
+      <SignUp03 />
+    </>
+  )
+}
