+++ pages/login/signup/04/index.js
@@ -1,0 +1,7 @@
+export default function Signup04Page(){
+  return(
+    <div>
+      
+    </div>
+  )
+}
