--- pages/login/signup/1/index.js
@@ -1,9 +1,0 @@
-import SignUpUI from "../../../../src/components/units/signup/signup1/Signup1.presenter";
-
-export default function Signup01(){
-  return(
-    <>
-      <SignUpUI />
-    </>
-  )
-}
