--- pages/login/signup/2/index.js
@@ -1,9 +1,0 @@
-import SignUpUI from "../../../../src/components/units/signup/signup2/Signup2.presenter";
-
-export default function Signup02(){
-  return(
-    <>
-      <SignUpUI />
-    </>
-  )
-}
