--- pages/login/signup/3/index.js
@@ -1,9 +1,0 @@
-import SignUpUI from "../../../../src/components/units/signup/signup3/Signup3.presenter";
-
-export default function Signup02(){
-  return(
-    <>
-      <SignUpUI />
-    </>
-  )
-}
