+++ src/components/units/loading/Loading.presenter.js
@@ -1,33 +1,10 @@
-import { useRouter } from "next/router";
 import * as S from "./Loading.styles";
 import Image from "next/image";
-import { useEffect, useState } from "react";
 
-export default function LoadingUI() {
-  const router = useRouter();
-
-  //3초뒤 자동으로 페이지 이동
-  useEffect(() => {
-    const timer = setTimeout(() => {
-      router.push('/login');
-    }, 1500);
-    //페이지 이동 후 타이며 초기화
-    return () => clearTimeout(timer);
-  }, []);
-
-  const [blurred, setBlurred] = useState(true);
-
-  useEffect(() => {
-    const timer = setTimeout(() => {
-      setBlurred(false);
-    }, 200);
-
-    return () => clearTimeout(timer);
-  }, []);
-
+export default function LoadingUI(props) {
   return (
     <S.WrapperLoading>
-      <S.BlurryElement blurred={blurred}>
+      <S.BlurryElement blurred={props.blurred}>
         <S.WrapperImage>
           <Image
             src="/images/loading/loading_paw.png"
