+++ src/components/units/signup/signup01/Signup01.styles.js
@@ -22,7 +22,7 @@ export const ProgressBarBlock = styled.div`
 `;
 
 export const ProgressBar = styled.progress`
-    width: 20%;
+    width: 25%;
     height: 3px;
     appearance: none;
