+++ styles/globals.css
@@ -74,12 +74,12 @@
 }
 */
 
-@font-face {
+/* @font-face {
   font-family: "PretenDard";
   src: url("/fonts/PretenDard-Regular.ttf") format("truetype");
   font-weight: 400;
   font-style: normal;
-}
+} */
 
 * {
   box-sizing: border-box;
